feat: add node n-api bindings
